test(cast): finish grouping tests by command
Move misplaced call, transaction, selector, Tempo, and replay tests to
their command modules. Keep Monad-only constants and helpers inside
the Monad test module.
