+++ crates/cast/tests/cli/abi.rs
@@ -17,21 +17,6 @@ casttest!(string_decode, |_prj, cmd| {
 "#]]);
 });
 
-casttest!(recover_authority, |_prj, cmd| {
-    let auth = r#"{
-        "chainId": "0x1",
-        "address": "0xb684710e6d5914ad6e64493de2a3c424cc43e970",
-        "nonce": "0x3dc1",
-        "yParity": "0x1",
-        "r": "0x2f15ba55009fcd3682cd0f9c9645dd94e616f9a969ba3f1a5a2d871f9fe0f2b4",
-        "s": "0x53c332a83312d0b17dd4c16eeb15b1ff5223398b14e0a55c70762e8f3972b7a5"
-    }"#;
-    cmd.args(["recover-authority", auth]).assert_success().stdout_eq(str![[r#"
-0x17816E9A858b161c3E37016D139cf618056CaCD4
-
-"#]]);
-});
-
 // Test cast abi-encode-event with indexed parameters
 casttest!(abi_encode_event_indexed, |_prj, cmd| {
     cmd.args([
@@ -238,40 +223,6 @@ casttest!(abi_encode_event_anonymous, |_prj, cmd| {
 [data]: 0x0000000000000000000000000000000000000000000000000000000000000020000000000000000000000000000000000000000000000000000000000000000568656c6c6f000000000000000000000000000000000000000000000000000000
 
 "#]]);
-});
-
-// https://github.com/foundry-rs/foundry/issues/11584
-// Tests that invalid hex calldata (odd length) produces a clear error message
-casttest!(cast_call_invalid_hex_calldata_error, |_prj, cmd| {
-    let rpc = next_rpc_endpoint(NamedChain::Mainnet);
-    cmd.args([
-        "call",
-        "0xdead000000000000000000000000000000000000",
-        "--data",
-        "0x0", // Invalid: odd length hex
-        "--rpc-url",
-        rpc.as_str(),
-    ])
-    .assert_failure()
-    .stderr_eq(str![[r#"
-Error: Invalid hex calldata '0x0': odd number of digits
-
-"#]]);
-});
-
-// https://github.com/foundry-rs/foundry/issues/11584
-// Tests that valid hex calldata works correctly
-casttest!(cast_call_valid_hex_calldata, |_prj, cmd| {
-    let rpc = next_rpc_endpoint(NamedChain::Mainnet);
-    cmd.args([
-        "call",
-        "0xdead000000000000000000000000000000000000",
-        "--data",
-        "0x00", // Valid: even length hex
-        "--rpc-url",
-        rpc.as_str(),
-    ])
-    .assert_success();
 });
 
 casttest!(abi_decode_output, |_prj, cmd| {
