+++ crates/cast/tests/cli/selectors.rs
@@ -314,3 +314,11 @@ true
 
 "#]]);
 });
+
+casttest!(signature_selector, |_prj, cmd| {
+    cmd.cast_fuse().args(["sig", "foo()"]).assert_success().stdout_eq("0xc2985578\n");
+    cmd.cast_fuse()
+        .args(["sig", "foo(address,uint256)"])
+        .assert_success()
+        .stdout_eq("0xbd0d639f\n");
+});
