+++ crates/cast/tests/cli/tempo.rs
@@ -1,7 +1,7 @@
 //! CLI tests for shared Tempo transaction options.
 
 use alloy_network::{ReceiptResponse, TransactionBuilder};
-use alloy_primitives::{Address, B256, U256, hex};
+use alloy_primitives::{Address, B256, U256, address, b256, hex, keccak256};
 use alloy_provider::Provider;
 use alloy_rpc_types::TransactionRequest;
 use alloy_serde::WithOtherFields;
@@ -11,7 +11,7 @@ use foundry_evm::core::tempo::PATH_USD_ADDRESS;
 use foundry_test_utils::util::OutputExt;
 use tempo_contracts::precompiles::{
     CURRENT_COMMITTEE_ADDRESS, ICurrentCommittee, IReceivePolicyGuard, ITIP20, ITIP403Registry,
-    TIP403_REGISTRY_ADDRESS,
+    TIP20_CHANNEL_RESERVE_ADDRESS, TIP403_REGISTRY_ADDRESS,
 };
 use tempo_hardfork::TempoHardfork;
 
@@ -1111,4 +1111,41 @@ casttest!(tip20_logo_set_validates_logo_uri_before_network_setup, |_prj, cmd| {
         .stderr_lossy();
 
     assert!(output.contains("client-side validation failed: InvalidLogoURI"), "got:\n{output}");
+});
+
+casttest!(channel_id_defaults, async |_prj, cmd| {
+    let (_api, handle) =
+        anvil::spawn(NodeConfig::test_tempo().with_hardfork(Some(TempoHardfork::T5.into()))).await;
+    let provider = handle.http_provider();
+    let chain_id = provider.get_chain_id().await.unwrap();
+
+    let payer = address!("0000000000000000000000000000000000000101");
+    let payee = address!("0000000000000000000000000000000000000202");
+    let salt = b256!("0000000000000000000000000000000000000000000000000000000000000042");
+    let expected = keccak256(
+        (
+            payer,
+            payee,
+            Address::ZERO,
+            PATH_USD_ADDRESS,
+            salt,
+            Address::ZERO,
+            B256::ZERO,
+            TIP20_CHANNEL_RESERVE_ADDRESS,
+            U256::from(chain_id),
+        )
+            .abi_encode(),
+    );
+
+    cmd.args([
+        "channel-id",
+        &payer.to_string(),
+        &payee.to_string(),
+        &PATH_USD_ADDRESS.to_string(),
+        &salt.to_string(),
+        "--rpc-url",
+        handle.http_endpoint().as_str(),
+    ])
+    .assert_success()
+    .stdout_eq(format!("{expected:#x}\n"));
 });
