+++ crates/cast/tests/cli/vaddr.rs
@@ -351,31 +351,4 @@ mod vaddr_e2e {
             );
         }
     );
-
-    // tests that displays a sample beacon block traces in Cancun
-    // https://github.com/foundry-rs/foundry/issues/12435
-    casttest!(test_beacon_block_root_in_cancun, |prj, cmd| {
-        prj.clear();
-        let eth_rpc_url = next_http_rpc_endpoint();
-        cmd.args([
-            "run",
-            "0xae290fe8c89c3e83dff20eeb2b8e3261bcdce0d66441c7056918dfb5fafe6d96",
-            "--rpc-url",
-            eth_rpc_url.as_str(),
-        ])
-        .assert_success()
-        .stdout_eq(str![[r#"
-Traces:
-  [45054] 0xB731392c0EB5BF2092f9f7B520DA551f70Ea9131::Claim{value: 46698476594582387}()
-    ├─ [4320] 0x000F3df6D732807Ef1319fB7B8bB8522d0Beac02::00000000(00000000000000000000000000000000000000000000000069091d4b) [staticcall]
-    │   └─ ← [Return] 0x70c7855161ec07af782df915fb3e81702df40f34972da3d740cdfc132ac926f6
-    ├─ emit NvStuck(param0: 0x6e6C36B970f8862bA3F148DEdAB8F98f5ed8b426, param1: 46698476594582387 [4.669e16], param2: 1762205003 [1.762e9])
-    └─ ← [Stop]
-
-
-Transaction successfully executed.
-[GAS]
-
-"#]]);
-    });
 }
