+++ crates/cast/tests/cli/wallet_keystore.rs
@@ -794,63 +794,3 @@ Password for keystore `testAccount` was changed successfully. [ADDRESS]
     let decrypted_private_key = B256::from_str(private_key_string).unwrap();
     assert_eq!(decrypted_private_key, test_private_key);
 });
-
-casttest!(mktx_tempo_access_key_uses_alloy_wallet, async |_prj, cmd| {
-    let (_, handle) = anvil::spawn(NodeConfig::test_tempo()).await;
-    let output = cmd
-        .args([
-            "mktx",
-            "0x0000000000000000000000000000000000000001",
-            "--rpc-url",
-            &handle.http_endpoint(),
-            "--chain",
-            "31337",
-            "--nonce",
-            "0",
-            "--gas-limit",
-            "100000",
-            "--gas-price",
-            "20000000000",
-            "--priority-gas-price",
-            "1000000000",
-            "--tempo.fee-token",
-            "0x20C0000000000000000000000000000000000000",
-            "--tempo.access-key",
-            "0x59c6995e998f97a5a0044966f0945389dc9e86dae88c7a8412f4603b6b78690d",
-            "--tempo.root-account",
-            "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266",
-        ])
-        .assert_success()
-        .get_output()
-        .clone();
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    let raw = hex::decode(stdout.trim().trim_start_matches("0x")).expect("decode raw transaction");
-    let TempoTxEnvelope::AA(signed) =
-        TempoTxEnvelope::decode_2718(&mut raw.as_slice()).expect("decode Tempo AA transaction")
-    else {
-        panic!("expected a Tempo AA transaction");
-    };
-    let TempoSignature::Keychain(signature) = signed.signature() else {
-        panic!("expected an account access-key signature");
-    };
-    assert_eq!(signature.user_address, address!("0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266"));
-    assert_eq!(signature.version, KeychainVersion::V2);
-    assert_eq!(
-        signature.key_id(&signed.tx().signature_hash()).unwrap(),
-        address!("0x70997970C51812dc3A010C7d01b50e0d17dc79C8")
-    );
-});
-
-casttest!(curl_call_rejects_browser_wallet, |_prj, cmd| {
-    let stderr = cmd
-        .args(["call", "0xdead000000000000000000000000000000000000", "--browser", "--curl"])
-        .assert_failure()
-        .get_output()
-        .stderr_lossy();
-
-    assert!(
-        stderr.contains("--browser cannot be combined with --curl; use --from <ADDRESS>"),
-        "unexpected stderr:\n{stderr}"
-    );
-});
